feat(UserServiceTests): refactor tests to use real DbContext and add InMemory database support
